requirements.txt
azure-storage-blob==12.*
boto3==1.*
ecodev-core==0.*
Fiona==1.8.22
GDAL==3.4.3
netCDF4==1.*
Shapely==2.*
typing-extensions==4.*
xlsxwriter==3.*
